Reorganize project folders and edit uri names
