--- index_new_maybe.js
@@ -1,5 +1,0 @@
-const app = require('./app');
-
-app.listen(3000, () => {
-  console.log('Running on port 3000');
-});
