+++ package.json
@@ -12,7 +12,7 @@
     "testStephen": "mocha -R min",
     "testUSE": "NODE_ENV=test mocha \"./test/**/*.test.js\" -R min",
     "testRealLatest": "export NODE_ENV=test || SET \"NODE_ENV=test\" && mocha \"./test/**/*.test.js\" -R min",
-    "test": "export NODE_ENV=test || SET \"NODE_ENV=test\" && mocha",
+    "test": "export NODE_ENV=test || SET \"NODE_ENV=test\" && mocha \"./server/test/**/*.test.js\"",
     "test-watch": "nodemon --exec \"npm test\""
   },
   "keywords": [],
