+++ server/routes/categories/index.js
