+++ server/scrapeScripts/helper.js
