--- test/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "rules": {
-    "handle-callback-err": 0
-  },
-  "env": {
-    "mocha": true
-  }
-}
